Added the MainMenu and rewrote how we store the access the ScreenMessage's physical addresses.

+++ Pins.h
@@ -10,16 +10,16 @@
 // ---- Screen pins ----
 
 // Set LOW to enable, HIGH to disable
-const int Pin_Screen_EnablePower = 5;
+const uint8_t Pin_Screen_EnablePower = 5;
 
-const int Pin_Screen_RS = A3;
-const int Pin_Screen_EN = A2;
-const int Pin_Screen_D4 = A1;
-const int Pin_Screen_D5 = A0;
-const int Pin_Screen_D6 = 6;
-const int Pin_Screen_D7 = 7;
+const uint8_t Pin_Screen_RS = A3;
+const uint8_t Pin_Screen_EN = A2;
+const uint8_t Pin_Screen_D4 = A1;
+const uint8_t Pin_Screen_D5 = A0;
+const uint8_t Pin_Screen_D6 = 6;
+const uint8_t Pin_Screen_D7 = 7;
 
 // ---- GPS pins ----
 
 // Set LOW to enable, HIGH to disable
-const int Pin_GPS_EnablePower = 4;
+const uint8_t Pin_GPS_EnablePower = 4;
